sor pages
A draft on SOR. While writing it down it became apparent, that we can simplify the interface by eliminating merging findRoute with buildSwap steps. Currently, finding a route and constructing a swap require knowledge of the exact in / out type. However, each of these actions uses a different term for this type.

Still todo: useBpts option, it's not covered on the SDK yet, probably makes sense to just add, because it's low level logic with Vault / Relayer encoding.

Otherwise the pages are now ready for review and I would appreciate your feedback.