Provide high level Vault API
It states that Token Accounting and Pool Logic are separated but it is not explained how. I think it is important to have at least high level understanding how exactly accounting is separated from pool logic. Maybe a diagram displaying all possible interactions between Vault and Pool can shed light on this.